move relevant tests to authorsMap test file
